+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_libs.prt_dict:  Print keys for nested dictionaries in the value.
 
 ### Changed
+- gen_libs.print_dict:  Added ability to add the dictionary to an email instance message.
 - gen_libs.del_not_and_list:  Replaced try/except with an intersect of two sets to capture items to be removed.
 - arg_parser.arg_wildcard:  Refactored the function for better functionality.
 - gen_libs.prt_dict:  Changed variable name to standard naming convention.